Create MixtureDistr_withCov
This code creates a mixture distribution (Bernoulli and Poisson) to deal with the zero-inflated dataset. The code also includes the addition of a covariate (for simplicity, the example is "month of year"). However, the model is NOT in a state space form where each timestep relies on the last timestep.